Serve real browsers the SPA on direct artist-page loads, not static HTML
/a/:slug and /artist/:slug had two renderers for one URL: the artist-page-static
edge function unconditionally served a hand-written static page (640px container,
no auth-aware header) for every direct load, while client-side navigation from
/artists rendered the real React ArtistPage (wide container, full Header with
Login). Same URL, two different UIs depending on how you got there.

Give the edge function the same crawler-only gate og-metadata.ts already uses at
"/": real browsers now fall through to the SPA (context.next()), so a direct load
and an in-app click render identically. Only social/indexing crawlers, which need
server-rendered content without waiting on a client fetch, still get the static
page. Extracted the shared bot-detection helpers to api/shared/crawler-detection.ts
so both edge functions use one implementation.
